Trying to train the network only once

+++ text_input.py
@@ -216,9 +216,10 @@ def main():
 #    reader = TextReader('./data/mr/', suffix_list=['neg', 'pos'])
     reader.prepare_data(vocab_size=4000, test_fraction=0.1)
 #    reader.prepare_data(vocab_size=15000, test_fraction=0.1)    
-    embedding = prepare_pretrained_embedding('./data/word2vec/GoogleNews-vectors-negative300.bin', reader.word2id)
-    # dump_to_file('./data/mr/emb.cPickle', embedding)
-    np.save('./data/mr/emb.npy', embedding)
+    if not os.path.isfile('./data/mr/emb.npy'):
+        embedding = prepare_pretrained_embedding('./data/word2vec/GoogleNews-vectors-negative300.bin', reader.word2id)
+        # dump_to_file('./data/mr/emb.cPickle', embedding)
+        np.save('./data/mr/emb.npy', embedding)
 
 
 if __name__ == '__main__':
